Allow user password length to be greater than 20
This is needed for a customer which requires passwords
greater than 20 characters.

For users requiring a minimum password length greater
than the 20 character limit, enable the ability to modify
the password after account creation by removing the user
from the IPMI group when updating to the 21+ char password.

This is a short term solution to solve the problem of allowing
for IPMI access but still allow the user to be dropped from IPMI,
removing the 20 char password limit, and allowing for a 21+ char
password. Ideally, Redfish and our backend would support some
sort of user groups and a user could drop the IPMI group/not
select the IPMI group at creation. Another solution might be
allowing the user to turn off IPMI support and removing the local
users from the IPMI user group.

- Adding an info tooltip to the create user modal to
  inform the ability to update password after account creation
- Add warning message when user password is updated to
  greater than 20
- Swap warning icon background image with <icon> directive

Signed-off-by: Yoshie Muranaka <yoshiemuranaka@gmail.com>
Change-Id: I4cb58f40cccafd8dfcd410fdff00104231ed1605
Signed-off-by: Gunnar Mills <gmills@us.ibm.com>